{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.pkg_name }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
authors = ["{{ cookiecutter.project_author }} <you@example.com>"]

[tool.poetry.dependencies]
python = "^{{ cookiecutter.python_version }}"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
